grsec: Apply grsecurity-2.9.1-3.2.44-201305132025.patch
commit 197968500d67cfd144658613250837810b109025
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon May 13 20:10:59 2013 -0400

    compile fix for PaX

 arch/x86/lib/usercopy_64.c |    4 ++--
 1 files changed, 2 insertions(+), 2 deletions(-)

commit a63fa11bbcb3f49cedd1057f9f7c7693f755a91a
Merge: 63253e2 1ef16b8
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon May 13 11:03:12 2013 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit 1ef16b89f336b44710a321b7afc7172873efaa82
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon May 13 11:01:58 2013 -0400

    Update to pax-linux-3.2.44-test105.patch:
    - another fix for kprobes under KERNEXEC
    - enabled user.pax.flags access for non-readable executables, fixes https://bugs.gentoo.org/show_bug.cgi?id=463170#c21
    - fixed some size overflow false positives, including the one in nfsd_cache_update
    - fixed sparc/constification compile error, reported by blake
    - UDEREF/amd64 should be a bit more efficient when disabled at boot time
    - fixed some unnecessary integer truncations that could trip up the size overflow plugin

 arch/sparc/kernel/us3_cpufreq.c     |    4 ++--
 arch/x86/ia32/ia32entry.S           |    4 ++--
 arch/x86/include/asm/pgtable.h      |    6 ++++--
 arch/x86/include/asm/uaccess.h      |    6 +++---
 arch/x86/kernel/kprobes.c           |    4 ++++
 arch/x86/lib/copy_user_nocache_64.S |    2 +-
 arch/x86/lib/getuser.S              |    8 ++++----
 arch/x86/lib/putuser.S              |    8 ++++----
 arch/x86/mm/fault.c                 |    6 +++---
 drivers/net/slip/slhc.c             |    2 +-
 drivers/staging/iio/ring_sw.c       |    2 +-
 fs/binfmt_elf.c                     |    6 +++---
 fs/nfsd/nfscache.c                  |    2 +-
 fs/xattr.c                          |   21 +++++++++++++++++++++
 include/linux/syscalls.h            |    2 +-
 include/linux/xattr.h               |    3 +++
 init/main.c                         |    3 +++
 kernel/futex_compat.c               |    2 +-
 net/socket.c                        |    2 +-
 19 files changed, 63 insertions(+), 30 deletions(-)

Signature-tree: ea9edba0784ded59391b424177fbcc77850dc343

+++ drivers/staging/iio/ring_sw.c
@@ -173,7 +173,7 @@ static int iio_read_first_n_sw_rb(struct iio_buffer *r,
 
 	u8 *initial_read_p, *initial_write_p, *current_read_p, *end_read_p;
 	u8 *data;
-	int ret, max_copied, bytes_to_rip, dead_offset;
+	long ret, max_copied, bytes_to_rip, dead_offset;
 
 	/* A userspace program has probably made an error if it tries to
 	 *  read something that is not a whole number of bpds.
